app/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
psycopg
psycopg2-binary
alembic
pydantic
pydantic-settings
sqlmodel
pika
aio-pika
replicate
openai
python-multipart
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
email-validator==2.1.0
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2